+++ app/src/main/java/com/practica/habitos/ui/components/AddHabit/ElegirCategoria.kt
@@ -1,4 +1,4 @@
-package com.practica.habitos.ui.theme.components.AddHabit
+package com.practica.habitos.ui.components.AddHabit
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
@@ -28,11 +28,11 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.compose.ui.window.Dialog
-import com.practica.habitos.Data.Models.Screens
+import com.practica.habitos.Domain.Models.Screens
 import com.practica.habitos.ui.theme.BackgroundHoyScree
 import com.practica.habitos.ui.theme.Rosadito
-import com.practica.habitos.ui.theme.Screen.AddHabitScreen.AddHabitViewModels
-import com.practica.habitos.ui.theme.components.categories.CategoryDisplay
+import com.practica.habitos.ui.screen.addHabitScreen.AddHabitViewModels
+import com.practica.habitos.ui.components.categories.CategoryDisplay
 
 @Composable
 fun ElegirCategoria(viewModels: AddHabitViewModels) {
