+++ app/src/main/java/com/practica/habitos/ui/components/navigationComponent/MenuLateral.kt
@@ -1,4 +1,4 @@
-package com.practica.habitos.ui.screen.components.navigationComponent
+package com.practica.habitos.ui.components.navigationComponent
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
