--- app/src/main/java/com/practica/habitos/ui/screen/components/AddHabit/ElegirDiasDeLaSemana.kt
@@ -1,3 +1,0 @@
-package com.practica.habitos.ui.screen.components.AddHabit
-
-//todo: componente para eligir los dias de la semana
